requirements.txt
alembic==1.9.2
aniso8601==9.0.1
click==8.1.3
Flask==2.2.2
Flask-JWT-Extended==4.4.4
flask-marshmallow==0.14.0
Flask-Migrate==4.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.0
PyJWT==2.6.0
pytz==2022.7.1
six==1.16.0
SQLAlchemy==1.4.46
Werkzeug==2.2.2
python-slugify==7.0.0
python-dotenv~=0.20.0
marshmallow-sqlalchemy==0.28.1